PR feedback. Refactored all the resizable code into a hook.
